requirements.txt
python-dotenv
azure-search-documents==11.6.0b4
azure-identity
openai
azure-ai-documentintelligence
tiktoken
werkzeug
tenacity
autogen-agentchat==0.4.4
autogen-core==0.4.4
autogen-ext[azure]
azure-monitor-opentelemetry-exporter
semantic-kernel
ipykernel
azure-ai-projects
pydantic